requirements.txt
celery
Flask
bcrypt==4.1.1
Flask-Bcrypt
Flask-Cors
Flask-JWT-Extended
flask-marshmallow
Flask-Migrate
flask-socketio
Flask-SQLAlchemy
Flask-PyMongo
flask-graphql
graphene
graphene-file-upload
graphene-sqlalchemy
marshmallow-sqlalchemy
psycopg2-binary
python-dotenv
python-slugify
redis